### 2023-09-04 #5798 @Kevin-Mattheus-Moerman
@brendanc490 thanks for this submission. I am the AEiC for this track and here to help process the initial steps. For the moment I've flagged this work for a scope review by our editorial board. In particular I need some help to see if this work conforms to what we class as research software.
The paper mentions that patterns created through the website could serve to help "approve devices for medical use". This could be an important application. However this functionality does not make it research software. Could you help provide some comments here as to how you think your software would be used in scientific research? It would help if what you describe paints a clear picture as to why this work is important to scientific research, i.e. such that one would feel the need to cite it in a subsequent paper. Utility tools are typically not cited, and do not require a paper. However software that has a direct impact on scientific findings would definitely need to be cited.
As you address this query, you may also decide to amend your paper and statement of need section on the research applications of this work.
Thanks.
