### 2023-09-14 #5798 @brendanc490
Hi Kevin,

This submission is coming on behalf of the FDA, so the tool is geared
towards regulatory science. We have developed a regulatory science tool
that enables the measurements of head-mounted display (HMD) quality by
creating customized scenes and patterns. This is in line with the
definition of regulatory science as it would fall under the creation of
tools for the assessment of the quality of FDA-related devices. This tool
in particular provides an easy way to display scenes on multiple headsets
so that we can assess and compare image quality. This is accomplished
through the use of known standardized patterns like the grille or dot array
and simple entities like circles, rectangles, or triangles. This
flexibility allows for the conduction of targeted examinations of specific
display characteristics. For instance, we can create and share packages
that are designed to assess display resolution, chromatic aberrations,
color mapping, and other spatiotemporal effects. This tool also addresses
the issue of pattern portability. Currently, HMDs are very heterogeneous,
making it difficult to deploy the same pattern to multiple headsets. This
barrier often comes up in standards meetings as an impeding force on the
development and validation of different HMDs. In short, this tool provides
scientists, medical device manufacturers, and regulators access to a
centralized scene to test various display devices and provide reports that
are easier to understand and, most importantly, replicate. Our goal was to
have a referenceable tool that we could use to guide manufacturers or
scientists when asking for specific image quality measurements on new VR/AR
devices.

Also, the paper has been updated to include this information. Let me know
if you have any further comments or questions.

Thanks,
Brendan

On Mon, Sep 4, 2023 at 11:55 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @brendanc490 <https://github.com/brendanc490> thanks for this submission.
> I am the AEiC for this track and here to help process the initial steps.
> For the moment I've flagged this work for a scope review by our editorial
> board. In particular I need some help to see if this work conforms to what
> we class as research software.
> The paper mentions that patterns created through the website could serve
> to help "approve devices for medical use". This could be an important
> application. However this functionality does not make it research software.
> Could you help provide some comments here as to how you think your software
> would be used in scientific research? It would help if what you describe
> paints a clear picture as to why this work is important to scientific
> research, i.e. such that one would feel the need to cite it in a subsequent
> paper. Utility tools are typically not cited, and do not require a paper.
> However software that has a direct impact on scientific findings would
> definitely need to be cited.
> As you address this query, you may also decide to amend your paper and
> statement of need section on the research applications of this work.
> Thanks.
>
