### 2023-10-02 #5798 @arfon
@brendanc490 – could you also explain how this repository is structured? I'm seeing a lot of JavaScript which I think is the bulk of the software contribution here?

On quick inspection, it's not clear how to test the software. Could you explain how that is accomplished here?
