### 2023-10-03 #5798 @brendanc490
I’d be happy to describe the repository. We submitted under the
JOSS-Submission branch. The tool we are submitting is the Custom Pattern
Creation Tool located in the "Custom" folder. It draws functionality from
the Compatibility tool to interface with virtual reality controllers and
headsets, which is also included in the "Compatibility" folder. The
majority of work was done in JavaScript as you pointed out, though there is
some HTML and CSS work. This encompasses everything except for the contents
of the "lib" folder which contains code libraries used in the project.

Upon opening the "Custom" folder you will see a readme containing detailed
steps on how to test the tool. We opted for manual tests since the tool is
focused on providing a user interface. The tests given highlight all the
core functionalities of the tool including creating, editing, saving, and
sharing a pattern.

Let me know if you have any questions.
Brendan
