### 2023-10-06 #5798 @codeling
First quick remark: It seems that the used license (CC0) is not OSI-approved.
