### 2023-10-06 #5798 @osorensen
@brendanc490, as pointed out by @codeling, the license CC0-1.0 is not OSI approved. Since this is a requirement for JOSS submissions, we cannot review this submission unless you change to an [OSI approved license](https://opensource.org/licenses/).
