### 2023-10-17 #5798 @brendanc490
We have decided to change the license to The Unlicense. This has been
updated in the repository.

On Mon, Oct 16, 2023 at 9:44 AM Øystein Sørensen ***@***.***>
wrote:

> @brendanc490 <https://github.com/brendanc490>, have you decided whether
> you will change the license or withdraw the publication?
>
