### 2023-11-05 #6021 @editorialbot
[REVIEW]: Open-Source Pattern Creation Tool for Medical Extended Reality Image Quality Assessment
**Submitting author:** @brendanc490 (<a href="http://orcid.org/0000-0002-1529-5900">Brendan Collins</a>)
**Repository:** https://github.com/DIDSR/WebXR-tools.git
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0
