### 2023-11-14 DIDSR/WebXR-tools#4 @codeling
README Improvement suggestions
In my review of [your JOSS submission](https://github.com/openjournals/joss-reviews/issues/6021), I noticed that the [review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) state that the README should contain a high-level of the software documentation; in that regard it is currently a bit lacking I think.

I would propose the following changes:
  - extend the introductory sentence a little, such as adding a summary of the functionality of the pattern creator, and mentioning all potential user groups (as mentioned in the paper).
  - add a link to the online version of the pattern creator (https://didsr.github.io/WebXR-tools/Custom/).
  - add a link to the  documentation (https://joss.readthedocs.io/en/latest/review_criteria.html).
  - add one sentence on possibility to run a local installation.
