### 2023-11-14 DIDSR/WebXR-tools#5 @codeling
Documentation - video format
In my review of [your JOSS submission](https://github.com/openjournals/joss-reviews/issues/6021), I noticed that the Demo video in the documentation (https://didsr.github.io/WebXR-tools/Custom/doc/introduction.html) doesn't play on my browser (Firefox 119 on both Ubuntu 22.04 and Windows 10); as far as I could gather, a webm container with h264 codec is generally not supported on Firefox (see https://bugzilla.mozilla.org/show_bug.cgi?id=1562862). Maybe the video could be provided in a different container/codec to enable a broader browser support?
