### 2023-11-16 #6021 @osorensen
Thanks @codeling. The license should definitely by the same also for the master branch. @brendanc490, can you please fix this?
