### 2023-11-17 DIDSR/WebXR-tools#6 @osorensen
License
This is related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6021). The LICENSE file in the main branch specifies that the license is CC0, which is not OSI approved.

https://github.com/DIDSR/WebXR-tools/blob/bf4271e612cdfcc54294662c5f720ef8874f3f4e/LICENSE#L1-L116
