### 2023-11-17 DIDSR/WebXR-tools#7 @osorensen
Review comments on paper
This is related to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6021).

Here are some suggestions for the paper:

- [x] Please refer to both figures in the main text. For example "Figure 1 shows how to edit a pattern using ... ."
- [x] On lines 36-37 you mention that the tool is built on A-Frame, Three.js, and WebXR. None of these tools have been described in scientific papers, and they don't have any suggested citation that I could find on their website. Still, I think it would be good to include a link to each of them. That way, users who are not familiar with these tools could easily click the links to find out more.
