### 2023-11-17 DIDSR/WebXR-tools#8 @osorensen
Screenshots of code in documentation
"TCA Test Instructions" contains several screenshots of code. In my opinion, this way of documenting is not sufficient to satisfy the review item "Functionality documentation". All users may not be able to read code very well. In addition, it will be hard to maintain, since you will have to manually take a new screenshot whenever the code is updated.

Please write out in plain English what the code lines are meant to tell.

https://github.com/openjournals/joss-reviews/issues/6021
