### 2023-11-17 DIDSR/WebXR-tools#9 @osorensen
Installation instructions
Another comment on my JOSS review. I realize that the code is mainly meant to be run through the website https://didsr.github.io/WebXR-tools/, but it should be possible to also run this locally, for users wishing to do so. Could you please add a section in the README containing installation instructions.

This relates to the review item shown below:

https://github.com/openjournals/joss-reviews/issues/6021
