### 2023-11-17 DIDSR/WebXR-tools#10 @osorensen
Automated tests
I could not find any automated tests for this software, nor any description of manual steps for confirming the validity.

Ideally, you could use continuous integration tests via [GitHub actions](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-nodejs).

This relates to the following review item:

https://github.com/openjournals/joss-reviews/issues/6021
