### 2023-11-17 DIDSR/WebXR-tools#11 @osorensen
Community guidelines
This relates to the following review item:

I suggest creating a file `CONTRIBUTING.md` in the top level of the repository, describing all of this.

https://github.com/openjournals/joss-reviews/issues/6021
