+++ Custom/README.md
@@ -1,168 +1,245 @@
-# Manual User Interface Tests #
+# Pattern Generation Tool #
+The Pattern Generation Tool provides a user interface for creating, editing, and displaying [WebXR](https://immersiveweb.dev/) scenes.  
+Scenes consist of 2 and 3 dimensional entities taken from [A-Frame](https://aframe.io/docs/1.4.0/core/entity.html).  
+These entities can be manipulated in a variety of different ways to create flexible patterns. 
+Scenes are grouped together in packages that can be downloaded and shared.
 
-Begin by opening this link https://didsr.github.io/WebXR-tools/Custom/
+### Features
+- Create packages of scenes that can be shared via a file or a link.
+- Order of scenes within the package can be easily changed.
+- Scenes are fully editable upon uploading or downloading.  
+- Scenes can be switched between using arrow keys or when in immersive mode, the buttons on the left and right controllers.
+- Default package available containing commonly used patterns:
+   - Red: A solid red background
+   - Green: A solid green background
+   - Blue: A solid blue background
+   - White: A solid white background
+   - Grille: A black and white background
+   - Crosshair: A white crosshair
+   - Line: A white vertical line
+   - Dot Array: A matrix of dots
+   - Circular Dot Array: A dot array arranged in a circle
+   - Checkerboard_w: A checkerboard with a white center tile
+   - Checkerboard_b: A checkerboard with a black center tile
+   - Ring_w1: A bullseye with a high pitch
+   - Ring_w2: A bullseye with a medium pitch
+   - Ring_w5: A bullseye with a low pitch
 
-## Test 1- Adding and editing a package
-This test will examine the package and pattern systems.  
-   1. Add a package with the name "Test 1" by pressing the *plus* icon next to **Packages**
-   2. Add three patterns to this package by pressing the *plus* icon next to **Pattern List**
-      - Name the first pattern "Pattern 1"
-      - Name the second pattern "Pattern 2"
-      - Name the third pattern "Pattern 3"
 
-   At this point, the tool should look like this:  
-   ![image](../Images/test1_1.PNG)
+# Directions For Use #
+### Packages
+1. [Create a Package](#createPackage)
+2. Edit a Package
+3. Display a Package
+4. Save a Package
+5. Share a Package
 
-   3. Reorder the patterns so that "Pattern 3" is the first pattern, "Pattern 2" is the second pattern, and "Pattern 1" is the last pattern
-      - Do this by dragging "Pattern 3" and dropping it in the first position
-      - Drag "Pattern 2" and drop it in the second position
-   4. Switch packages to the "default" package
-   5. Multi-select the patterns "crosshair" and "ring_w2" by holding ctrl while clicking on the patterns
-   6. Copy these patterns by pressing Ctrl+c
-   7. Switch packages to the "Test 1 Package" package
-   8. Paste the copied patterns by pressing Ctrl+v
+### Patterns
+1. Create a Pattern
+2. Edit a Pattern
+3. Display a Pattern
+4. Save a Pattern
+5. Share a Pattern
 
-   At this point, the tool should look like this:
-   ![image](../Images/test1_2.PNG)
 
-## Test 2- Adding and editing a pattern
-This test will examine the pattern and entity systems.  
-   9. Rename the "crosshair" pattern to "Pattern 4"
-       - Select the "crosshair" pattern by clicking on it
-       - Open settings by pressing the *gear* icon at the top right of the tab
-       - Under **Pattern Settings** select the *Rename* button
-       - Enter "Pattern 4" in the textbox and press OK
-   10. Delete the "ring_w2" pattern by pressing the *trash* icon next to **Pattern List**
-       - Select OK on the confirmation
[…]
