### 2023-12-07 DIDSR/WebXR-tools#4 @brendanc490
These are great points. I’ve updated the README to contain the links and a broad overview of the tool and its use case. I have also added a section about how to run the tool locally. https://github.com/DIDSR/WebXR-tools/blob/JOSS-Submission/README.md
