### 2023-12-07 DIDSR/WebXR-tools#5 @brendanc490
Thanks for pointing this out. I have converted the file to an mp4, which should be more widely compatible. Please check that the video plays on your setup. https://didsr.github.io/WebXR-tools/Custom/doc/introduction.html#demo
