### 2023-12-07 DIDSR/WebXR-tools#6 @brendanc490
Thanks for pointing this out. I have updated the license for the main branch. https://github.com/DIDSR/WebXR-tools/blob/main/LICENSE
