### 2023-12-07 DIDSR/WebXR-tools#7 @brendanc490
Thanks for pointing this out. I have added references to my figures within the paper as well as links to the open-source libraries used. You can check the new version of the manuscript.
