### 2023-12-07 DIDSR/WebXR-tools#8 @brendanc490
I can see how this would be confusing. The contents in the main branch contain tools that I do not wish to submit, including this TCA test. Only contents of the JOSS submission branch are included in the submission. I can remove this from the main branch if it seems confusing.
