### 2023-12-07 DIDSR/WebXR-tools#9 @brendanc490
This is a good point. I will add the instructions on how to run this tool locally. The steps here are to download the distribution and launch the index.html locally. I added them to the README. https://github.com/DIDSR/WebXR-tools/blob/JOSS-Submission/README.md
