### 2023-12-07 DIDSR/WebXR-tools#10 @brendanc490
The testing procedures were not easy to find, I apologize. Here is the link to the tests: https://github.com/DIDSR/WebXR-tools/blob/JOSS-Submission/Custom/README.md. I have opted for manual testing in this case because the tool itself is a user interface that heavily relies on human interaction. The linked README file contains a series of tests that can be walked through. I wanted to ensure that task completion is easy and understandable.
