### 2023-12-08 DIDSR/WebXR-tools#11 @brendanc490
Thanks for pointing this out. We will use the issues feature on GitHub to allow for reporting bugs, errors, and asking questions. I added this to the documentation https://didsr.github.io/WebXR-tools/Custom/doc/contributing.html.
