+++ Custom/doc/_sources/index.rst.txt
@@ -16,7 +16,6 @@ The Pattern Generation Tool provides a user interface for creating, editing, and
    patterns
    entities
    extra
-   contributing
 
 
 
