### 2023-12-11 DIDSR/WebXR-tools#9 @osorensen
Thanks, but the installation instructions for running locally did not work for me. Here is what I did:

## Step 1

I am on the `JOSS-Submission` branch, and run `open ./index.html`:

## Step 2

I get to the frontpage:

## Step 3

Clicking into any of the two options simply gives me a listing of the files in the directory

Do I need anything else installed?
