### 2023-12-11 DIDSR/WebXR-tools#10 @osorensen
Thanks, I went through most of the procedure and verified that it works. Some minor comments to the instructions:

- I think correct spelling for the control tab is "Ctrl", with capital "C".
- On Mac, I had to use the command tab wherever you mention the control tab, even though there is a control tab on Mac keyboards as well. Perhaps at first mention of "Ctrl", you could add in parentheses something like "(alternatively Command on Mac)".
