### 2023-12-11 DIDSR/WebXR-tools#5 @codeling
The video plays now on Firefox 👍 .
