### 2023-12-21 DIDSR/WebXR-tools#10 @brendanc490
Thanks for pointing this out. I changed the spelling and capitalization of the Ctrl instructions as well as added information about how to complete them on Mac.
