### 2023-12-21 DIDSR/WebXR-tools#9 @brendanc490
There was an error in the first index.html, thanks for pointing this out. I have updated that index.html so that it should now always open the tools, not their directory.
