### 2023-12-25 DIDSR/WebXR-tools#4 @codeling
Looks good.
I guess JOSS-Submission branch will soon be merged into main?
