### 2024-01-01 #6021 @osorensen
👋 @brendanc490, I wanted to raise the issue that you don't want the main branch of the repository to be part of the review. This is of relevance, e.g., [here](https://github.com/DIDSR/WebXR-tools/issues/8). It is absolutely fine that you use the JOSS review branch for development during the review, but unless you have very compelling arguments, I think the final revised version should be merged onto the main branch, so that the reviewed software becomes the default. Does that sound ok to you?
