### 2024-01-02 #6021 @brendanc490
Thanks for pointing this out, you're absolutely right. We will merge the
JOSS review branch onto the main branch when the review is complete.

On Mon, Jan 1, 2024 at 12:04 PM Øystein Sørensen ***@***.***>
wrote:

> 👋 @brendanc490 <https://github.com/brendanc490>, I wanted to raise the
> issue that you don't want the main branch of the repository to be part of
> the review. This is of relevance, e.g., here
> <https://github.com/DIDSR/WebXR-tools/issues/8>. It is absolutely fine
> that you use the JOSS review branch for development during the review, but
> unless you have very compelling arguments, I think the final revised
> version should be merged onto the main branch, so that the reviewed
> software becomes the default. Does that sound ok to you?
>
