### 2024-01-02 DIDSR/WebXR-tools#4 @brendanc490
Yes, that's correct. When the review is complete we will merge this branch into main.
