### 2024-01-04 DIDSR/WebXR-tools#8 @osorensen
Thanks, then I assume this will be fixed once the JOSS branch gets merged onto the main branch.
