### 2024-01-04 #6021 @osorensen
👋 @codeling, could you please update [your checklist](https://github.com/openjournals/joss-reviews/issues/6021#issuecomment-1794329716) based on the changes made by the authors? If there are further issues that should be addressed, please let us know.
