### 2024-01-05 #6021 @osorensen
@brendanc490, could you please complete the tasks under the header "Additional Author Tasks After Review is Complete" in [this checklist](https://github.com/openjournals/joss-reviews/issues/6021#issuecomment-1878383140), and report back here when done?

In addition, please merge the JOSS branch onto the main branch now that the review is complete.
