### 2024-01-05 #6021 @brendanc490
I have completed the tasks in the checklist. The DOI from Zenodo is: [10.5281/zenodo.10463335](https://zenodo.org/doi/10.5281/zenodo.10463335). Also, the JOSS branch has been merged onto main. Please let me know if I missed anything.
