### 2024-01-11 #6021 @Kevin-Mattheus-Moerman
@brendanc490 I am the AEiC and here to help with final steps to process acceptance in JOSS. I have check this review, the archive link, your repository, and the paper. Most seems in order. I only have the below points that requires your attention:

- [ ] For your affiliations please add city, state, and country information. And please spell out the country name, e.g. Unites States of America.
- [ ] The sentence `This is a contribution of the U.S. Food and Drug Administration and not subject to copyright`, seems to lack a full stop at the end.
